# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
      "ApplicationRun": {
        "additionalProperties": false,
        "properties": {
          "autoSubmitAuthorized": {
            "default": false,
            "title": "Autosubmitauthorized",
            "type": "boolean"
          },
          "candidateProfileId": {
            "format": "uuid",
            "title": "Candidateprofileid",
            "type": "string"
          },
          "candidateProfileVersion": {
            "minimum": 1.0,
            "title": "Candidateprofileversion",
            "type": "integer"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "jobId": {
            "format": "uuid",
            "title": "Jobid",
            "type": "string"
          },
          "jobUrl": {
            "format": "uri",
            "minLength": 1,
            "title": "Joburl",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/WorkflowState",
            "default": "created"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": ["jobId", "candidateProfileId", "candidateProfileVersion", "jobUrl"],
        "title": "ApplicationRun",
        "type": "object"
      },
      "Body_import_resume_v1_profile_resume_post": {
        "properties": {
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": ["file", "expected_version"],
        "title": "Body_import_resume_v1_profile_resume_post",
        "type": "object"
      },
...
      "CandidateProfileSnapshot": {
        "additionalProperties": false,
        "properties": {
          "profile": {
            "$ref": "#/components/schemas/CandidateProfile"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": ["profile"],
        "title": "CandidateProfileSnapshot",
        "type": "object"
      },
      "CreateRunRequest": {
        "additionalProperties": false,
        "properties": {
          "autoSubmitAuthorized": {
            "default": false,
            "title": "Autosubmitauthorized",
            "type": "boolean"
          },
          "candidateProfileId": {
            "format": "uuid",
            "title": "Candidateprofileid",
            "type": "string"
          },
          "candidateProfileVersion": {
            "minimum": 1.0,
            "title": "Candidateprofileversion",
            "type": "integer"
          },
          "jobId": {
            "format": "uuid",
            "title": "Jobid",
            "type": "string"
          },
          "jobUrl": {
            "format": "uri",
            "minLength": 1,
            "title": "Joburl",
            "type": "string"
          },
          "startImmediately": {
            "default": true,
            "title": "Startimmediately",
            "type": "boolean"
          }
        },
        "required": ["jobId", "candidateProfileId", "candidateProfileVersion", "jobUrl"],
        "title": "CreateRunRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ResumeImportResult": {
        "additionalProperties": false,
        "properties": {
          "document": {
            "$ref": "#/components/schemas/SourceDocument"
          },
          "snapshot": {
            "$ref": "#/components/schemas/CandidateProfileSnapshot"
          }
        },
        "required": ["snapshot", "document"],
        "title": "ResumeImportResult",
        "type": "object"
      },
      "SaveCandidateProfileRequest": {
        "additionalProperties": false,
        "properties": {
          "expectedVersion": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expectedversion"
          },
          "profile": {
            "$ref": "#/components/schemas/CandidateProfileInput"
          }
        },
        "required": ["profile"],
        "title": "SaveCandidateProfileRequest",
        "type": "object"
      },
...
      "SetEvidenceVerificationRequest": {
        "additionalProperties": false,
        "properties": {
          "expectedVersion": {
            "minimum": 1.0,
            "title": "Expectedversion",
            "type": "integer"
          },
          "verified": {
            "title": "Verified",
            "type": "boolean"
          }
        },
        "required": ["expectedVersion", "verified"],
        "title": "SetEvidenceVerificationRequest",
        "type": "object"
      },
...
      "WorkflowEvent": {
        "additionalProperties": false,
        "properties": {
          "eventType": {
            "title": "Eventtype",
            "type": "string"
          },
          "fromState": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkflowState"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "occurredAt": {
            "format": "date-time",
            "title": "Occurredat",
            "type": "string"
          },
          "reasonCode": {
            "title": "Reasoncode",
            "type": "string"
          },
          "runId": {
            "format": "uuid",
            "title": "Runid",
            "type": "string"
          },
          "safeDetails": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "number"
                },
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Safedetails",
            "type": "object"
          },
          "sequence": {
            "minimum": 1.0,
            "title": "Sequence",
            "type": "integer"
          },
          "toState": {
            "$ref": "#/components/schemas/WorkflowState"
          }
        },
        "required": ["runId", "sequence", "eventType", "toState", "reasonCode"],
        "title": "WorkflowEvent",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "CareerFlow Local API",
    "version": "0.1.5"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/events": {
      "get": {
        "operationId": "events_v1_events_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Events"
      }
    },
...
    "/v1/profile": {
      "get": {
        "operationId": "get_profile_v1_profile_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/CandidateProfileSnapshot"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Profile V1 Profile Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Profile"
      },
      "put": {
        "operationId": "save_profile_v1_profile_put",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveCandidateProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateProfileSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Profile"
      }
    },
    "/v1/profile/evidence/{evidence_id}/verification": {
      "put": {
        "operationId": "set_evidence_verification_v1_profile_evidence__evidence_id__verification_put",
        "parameters": [
          {
            "in": "path",
            "name": "evidence_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Evidence Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetEvidenceVerificationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateProfileSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Evidence Verification"
      }
    },
    "/v1/profile/resume": {
      "post": {
        "operationId": "import_resume_v1_profile_resume_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_resume_v1_profile_resume_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeImportResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Resume"
      }
    },
    "/v1/runs": {
      "get": {
        "operationId": "list_runs_v1_runs_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ApplicationRun"
                  },
                  "title": "Response List Runs V1 Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Runs"
      },
      "post": {
        "operationId": "create_run_v1_runs_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationRun"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Run"
      }
    },
    "/v1/runs/{run_id}/events": {
      "get": {
        "operationId": "list_run_events_v1_runs__run_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WorkflowEvent"
                  },
                  "title": "Response List Run Events V1 Runs  Run Id  Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Run Events"
      }
    },
    "/v1/runs/{run_id}/transitions": {
      "post": {
        "operationId": "transition_run_v1_runs__run_id__transitions_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowEvent"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Transition Run"
      }
    }
  }
}